Flask_API/requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
Bottleneck==1.2.1
captcha==0.3
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
colorama==0.4.1
colour==0.1.5
cycler==0.10.0
cymem==2.0.2
Cython==0.29.10
dask==1.2.2
dataclasses==0.6
DateTime==4.3
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
dlib==19.17.0
docutils==0.15.2
easydict==1.9
entrypoints==0.3
faiss==1.5.3
faiss-cpu==1.6.3
fastai==1.0.53.post2
fastprogress==0.1.21
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
gast==0.2.2
gensim==3.8.1
gevent==1.4.0
google-pasta==0.1.7
googleapis-common-protos==1.6.0
graphviz==0.11
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.6
imageio==2.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
jieba==0.39
Jinja2==2.10.1
jmespath==0.9.5
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-bert==0.81.0
keras-embed-sim==0.7.0
keras-layer-normalization==0.14.0
keras-multi-head==0.22.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.0
keras-self-attention==0.41.0
keras-transformer==0.32.0
kiwisolver==1.1.0
lmdb==0.95
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.0
numexpr==2.6.9
numpy==1.16.1
nvidia-ml-py3==7.352.0
opencv-python==4.1.0.25
openslide-python==1.1.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plac==0.9.6
pprint==0.1
preshed==2.0.1
prometheus-client==0.7.0
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.3
ptyprocess==0.6.0
pycairo==1.16.2
pycparser==2.19
pycurl==7.43.0.2
pydot==1.4.1
Pygments==2.4.2
pygobject==3.26.1
pymongo==3.8.0
pyparsing==2.4.0
pyrsistent==0.15.2
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pywinpty==0.5.5
pyxattr==0.6.0
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.1
redis==3.2.1
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.3.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
scour==0.36
seaborn==0.9.0
Send2Trash==1.5.0
shadowsocks==3.0.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.11.1
soupsieve==1.9.1
spacy==2.1.4
sqlparse==0.3.0
srsly==0.0.6
ssh-import-id==5.7
systemd-python==234
tb-nightly==1.14.0a20190301
tensorboardX==2.0
tensorflow-gpu==2.0.0a0
tensorflow-metadata==0.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
tfds-nightly==1.0.2.dev201904090105
thinc==7.0.4
toolz==0.9.0
torch==1.1.0
torchvision==0.3.0
tornado==5.1.1
tqdm==4.32.1
traitlets==4.3.2
typing==3.6.6
unattended-upgrades==0.1
urllib3==1.22
utils==0.9.0
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wincertstore==0.2
wordcloud==1.5.0
wrapt==1.11.1
xgboost==0.90
xlrd==1.2.0
yacs==0.1.6
youtube-dl==2018.3.14
zope.interface==4.6.0
